server/requirements.txt
django>=4.0
paho-mqtt>=1.6.1
gunicorn>=20.1.0

supabase
